refactor: move AverageNumber to error_recorder
